requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
python-dotenv==1.0.0
yookassa==3.2.0
aiohttp==3.9.1
